requirements.txt
streamlit==1.56.0
gspread==6.2.1
google-auth==2.48.0
pandas==3.0.0
